add Upstart 1.8 job file
::Release Notes:
This fix is simply moving the Upstart 1.8 job file for luna-sysservice
from the starfish-initscripts package to its proper home.

:Detailed Notes:
As mentioned above, this fix is just moving the upstart script - in the
initial implementation of Upstart 1.8, many 'component' upstart jobs
were modified for 1.8 and packaged with starfish-initscripts to ease
the transition from Upstart 0.3.11 to 1.8.  The plan was and it to
move the 'imported' scripts back to their home package as soon as is
reasonable.
Note: for this change to build the fix for [GF-8508] needs to be in
luna-sysservice also.  (affects CMakeLists.txt)

:Testing Performed:
Since the file itself is functionally unchanged from that present in
starfish-initscripts, the testing done running Upstart 1.8 + s-i over
the last couple of months still applies.

The testing done specifically for this fix is to locally build the
component and inspect the output directory:
starfish/BUILD-goldfinger/work/goldfinger-starfish-linux-gnueabi/ \
   /luna-sysservice/2.0.1-3-r5starfish3/image/etc/init and ..../etc/event.d
to confirm that the properly configured upstart scripts end up in their
respective locations.

:QA Notes:

:Issues Addressed:
[GF-5620]  update coreOS upstart scripts for 1.8

Open-webOS-DCO-1.0-Signed-off-by: Ed Chejlava <ed.chejlava@lge.com>

Change-Id: I23d914d353e1455779c0d4d15f9714e0534befb3
Reviewed-on: https://g2g.palm.com/2798
Reviewed-by: Build Verification
Reviewed-by: Ed Chejlava <ed.chejlava@lge.com>
Tested-by: Ed Chejlava <ed.chejlava@lge.com>
Reviewed-by: Roger Stringer <roger.stringer@lge.com>